time_and_memory/runit.c

+++ time_and_memory/ps_problem_sieve.c
@@ -2,11 +2,11 @@
 
 void fill_sieve(struct sieve_t* sv)
 {
-    for (int long long i = 2; i <= sv->n; ++i)
+    for (int i = 2; i <= sv->n; ++i)
     {
         if (sv->s[i] == 0)
         {
-            int long long index = i * i;
+            int index = i * i;
 
             while (index <= sv->n)
             {
@@ -18,10 +18,10 @@ void fill_sieve(struct sieve_t* sv)
 }
 
 
-int long long nth_prime(struct sieve_t* sv, int long long N)
+int nth_prime(struct sieve_t* sv, int N)
 {
-    int long long counter = 0;
-    for (int long long i = 2; i <= sv->n; ++i)
+    int counter = 0;
+    for (int i = 2; i <= sv->n; ++i)
     {
         if (sv->s[i] == 0) ++counter;
